requirements.txt
slackclient>=2.2,<2.3